+++ include/nstd/net/tcp.h
@@ -1,10 +1,10 @@
 #ifndef NSTD_NET_TCP_H_INCLUDED
 #define NSTD_NET_TCP_H_INCLUDED
-#include "../collections/vec.h"
 #include "../core/def.h"
 #include "../core/slice.h"
 #include "../core/str.h"
 #include "../nstd.h"
+#include "../vec.h"
 NSTDCPPSTART
 
 /// Represents a TCP server.