+++ include/nstd/net/udp.h
@@ -1,11 +1,11 @@
 #ifndef NSTD_NET_UDP_H_INCLUDED
 #define NSTD_NET_UDP_H_INCLUDED
-#include "../collections/vec.h"
 #include "../core/def.h"
 #include "../core/slice.h"
 #include "../core/str.h"
 #include "../nstd.h"
 #include "../string.h"
+#include "../vec.h"
 NSTDCPPSTART
 
 /// Represents a UDP socket.