+++ src/collections.rs
@@ -1,4 +1,3 @@
 pub mod bit_mask;
 pub mod rc;
 pub mod stack;
-pub mod vec;