+++ src/net/tcp.rs
@@ -1,6 +1,6 @@
 use crate::{
-    collections::vec::NSTDVec,
     core::{def::NSTDErrorCode, slice::NSTDSlice, str::NSTDStr},
+    vec::NSTDVec,
 };
 use std::{
     io::{BufRead, BufReader, Write},
@@ -94,7 +94,7 @@ pub unsafe extern "C" fn nstd_net_tcp_stream_read(stream: NSTDTCPStream) -> NSTD
         Ok(bytes) => NSTDVec::from(bytes),
         _ => {
             let null_slice = crate::core::slice::nstd_core_slice_new(0, 0, std::ptr::null_mut());
-            crate::collections::vec::nstd_collections_vec_from_existing(0, &null_slice)
+            crate::vec::nstd_vec_from_existing(0, &null_slice)
         }
     }
 }